[权重策略]支持权重的抽奖装配

+++ big-market-code/big-market-code-infrastructure/src/main/java/com/chige/infrastructure/persistent/po/StrategyRule.java
@@ -17,12 +17,12 @@ public class StrategyRule {
     /**
      * 抽奖策略ID
      */
-    private Integer strategyId;
+    private Long strategyId;
 
     /**
      * 抽奖奖品ID【规则类型为策略，则不需要奖品ID】
      */
-    private Integer awardId;
+    private Long awardId;
 
     /**
      * 抽象规则类型；1-策略规则、2-奖品规则
